requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9,<3.0
Pillow>=9.0,<11.0
django-cors-headers>=4.0,<5.0
django-environ>=0.10,<0.11